requirements.txt
pypiwin32
pywebview
httpx
Flask
pyclip
pycryptodome